Overrode equals in Piece and BoardState. Restructured test location
